Move the addtrips page to the trips tab
Additionally, removed the add trip button on the home page for now since
the main functionality of the home page is not to add a trip
