+++ app/screens/(tabs)/trips/_layout.tsx
@@ -4,6 +4,7 @@ export default function Layout()  {
     return (
         <Stack>
             <Stack.Screen name="index" options={{ headerShown: false }} />
+            <Stack.Screen name="components/addTrip" options={{ headerShown: false }} />
         </Stack>
     );
 }
