+++ app/screens/(tabs)/trips/components/addTrip.tsx
